Some edits to VM deployment and adding AMI output

+++ scripts/cleanup.sh
@@ -65,7 +65,9 @@ echo "		http://\$ip:8080				       "
 echo "*********************************************************************************"
 echo;echo
 EOF
-else
+
+elif [[ $BUILD_NAME = "vbox" ]]; then
+
 echo "BUILD for VBOX"
 cat > /home/gpadmin/start_all.sh << EOF
 echo "*********************************************************************************"
@@ -100,6 +102,41 @@ echo "**************************************************************************
 echo;echo
 EOF
 
+else 
+
+echo "BUILD for OTHERS"
+cat > /home/gpadmin/start_all.sh << EOF
+echo "*********************************************************************************"
+echo "* Script starts the Greenplum DB, Greenplum Control Center, and Apache Zeppelin *"
+echo "*********************************************************************************"
+echo "* Starting Greenplum Database..."
+source /usr/local/greenplum-db/greenplum_path.sh
+source /usr/local/greenplum-cc-web/gpcc_path.sh
+source /home/gpadmin/gp-wlm/gp-wlm_path.sh
+export MASTER_DATA_DIRECTORY=/gpdata/master/gpseg-1
+gpstart -a
+echo "* Greenplum Database Started."
+echo "* Starting Greenplum Command Center..."
+gpcmdr --start
+echo "* Greenplum Command Center Started."
+echo "* Starting Apache Zeppelin Server...."
+sudo /usr/local/$ZEPPELIN_VERSION/bin/zeppelin-daemon.sh start
+echo "* Apache Zeppelin Server Started."
+echo "*********************************************************************************"
+echo "* Updating Tutorial Files..."
+cd ~/gpdb-sandbox-tutorials;git pull > /dev/null 2>&1;tar xvfz faa.tar.gz;cd
+echo "* Tutorials Updated."
+echo "*********************************************************************************"
+echo " Pivotal Greenplum Database Started on port 5432        "
+echo " Pivotal Greenplum Command Center started on port 28080 "
+echo "          http://localhost:28080                              "
+echo "          Username: gpmon                                "
+echo "          Password: pivotal                              "
+echo " Apache Zeppelin started on port 8080                    "
+echo "          http://localhost:8080                                       "
+echo "*********************************************************************************"
+echo;echo
+EOF
 
 fi
 
@@ -144,6 +181,7 @@ rm -rf /tmp/bins
 rm -rf /tmp/configs
 
 # Defragment the blocks or else the generated VM image will still be huge
-
+if [[ $BUILD_NAME = "vbox" || $BUILD_NAME = "vmware" ]]; then
 dd if=/dev/zero of=/bigemptyfile bs=4096k
 rm -rf /bigemptyfile
+fi
